+++ src/Felicity.IntegrationTests/ToManyRelationshipPatchResourceAndSelf.fs
@@ -613,10 +613,10 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/children/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child1' and ID 'c1' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/relationships/children/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }
@@ -644,10 +644,10 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/children/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child2' and ID 'c22' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/relationships/children/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }
@@ -699,10 +699,10 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/otherChildren/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child2' and ID 'c22' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/relationships/otherChildren/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }
@@ -1310,10 +1310,10 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child1' and ID 'c1' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }
@@ -1332,10 +1332,10 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child2' and ID 'c22' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }
@@ -1369,10 +1369,10 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/0" @>
       test <@ json |> getPath "errors[1].status" = "404" @>
-      test <@ json |> getPath "errors[1].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[1].detail" = "The related resource with type 'child2' and ID 'c22' does not exist" @>
       test <@ json |> getPath "errors[1].source.pointer" = "/data/1" @>
       test <@ json |> hasNoPath "errors[2]" @>
     }