+++ src/Felicity.IntegrationTests/ToOneNullableRelationshipPatchResourceAndSelf.fs
@@ -564,7 +564,7 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/child/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }
@@ -588,7 +588,7 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/child/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }
@@ -636,7 +636,7 @@ let tests1 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/otherChild/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }
@@ -1244,7 +1244,7 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }
@@ -1264,7 +1264,7 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }
@@ -1304,7 +1304,7 @@ let tests2 =
       response |> testStatusCode 404
       let! json = response |> Response.readBodyAsString
       test <@ json |> getPath "errors[0].status" = "404" @>
-      test <@ json |> getPath "errors[0].detail" = "The related resource does not exist" @>
+      test <@ json |> getPath "errors[0].detail" = "The related resource with type 'child2' and ID 'c2' does not exist" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
     }